refactor: Corrige schema do componente e suas propriedades
