+++ src/components/ui-icon/ui-icon-enum.ts
@@ -2,4 +2,3 @@ export enum iconEnum {
   arrowLeft = 'arrow-left',
   arrowRight = 'arrow-right'
 }
-
